Changed FanClicker for FanInteractions and moved all components related to interactions there
