+++ Boccia-Unity/Assets/Scenes/Demo Fan Scene 1.unity
@@ -1710,6 +1710,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   ball: {fileID: 1616241089}
+  dropBar: {fileID: 0}
 --- !u!1001 &713752003
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1749,6 +1750,14 @@ PrefabInstance:
     - target: {fileID: -6058800571561212690, guid: ac665ac115a02a344af5c3955a46d709, type: 3}
       propertyPath: rowSpacing
       value: 0.1
+      objectReference: {fileID: 0}
+    - target: {fileID: -6058800571561212690, guid: ac665ac115a02a344af5c3955a46d709, type: 3}
+      propertyPath: _innerRadius
+      value: 4
+      objectReference: {fileID: 0}
+    - target: {fileID: -6058800571561212690, guid: ac665ac115a02a344af5c3955a46d709, type: 3}
+      propertyPath: _outerRadius
+      value: 5
       objectReference: {fileID: 0}
     - target: {fileID: -6058800571561212690, guid: ac665ac115a02a344af5c3955a46d709, type: 3}
       propertyPath: columnSpacing
@@ -1804,7 +1813,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 7462433658856035897, guid: ac665ac115a02a344af5c3955a46d709, type: 3}
       propertyPath: m_Name
-      value: Fan Prefab
+      value: RampControlFan
       objectReference: {fileID: 0}
     - target: {fileID: 7650961842829309085, guid: ac665ac115a02a344af5c3955a46d709, type: 3}
       propertyPath: r1
